+++ Thrower/Validation/ValidationError.cs
@@ -21,11 +21,14 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT
 // OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System;
+
 namespace PommaLabs.Thrower.Validation
 {
     /// <summary>
     ///   Represents an error found while validating an object.
     /// </summary>
+    [Serializable]
     public struct ValidationError
     {
         /// <summary>